Move store to models
