+++ src/models/experts/index.ts
@@ -1,9 +1,5 @@
 /* eslint-disable import/no-cycle */
-export {
-  fetchExperts,
-  fetchExpertById,
-  fetchExpertMaterials,
-} from './asyncActions';
+export { fetchExperts, fetchExpertMaterials } from './asyncActions';
 export { resetMaterials, loadPosts, expertsReducer } from './reducers';
 export {
   selectExperts,
