+++ src/models/rootReducer.ts
@@ -2,7 +2,6 @@ import { combineReducers } from 'redux';
 import { expertsReducer } from './experts';
 import { postCreationReducer } from './postCreation';
 import { propertiesReducer } from './properties';
-import dataReducer from './dataSlice';
 import { signInReducer } from './user';
 import { materialsReducer } from './materials';
 import { mainReducer } from './main';
@@ -13,7 +12,6 @@ const rootReducer = combineReducers({
   properties: propertiesReducer,
   experts: expertsReducer,
   newPostDraft: postCreationReducer,
-  data: dataReducer,
   currentUser: signInReducer,
 });
