+++ src/old/modules/posts/components/PostViewWrapper.tsx
@@ -7,7 +7,7 @@ import PostView from './PostView';
 import { sanitizeHtml } from '../../../lib/utilities/sanitizeHtml';
 import { PageTitle } from '../../../lib/components/Pages/PageTitle';
 import { setGALocation } from '../../../../utilities/setGALocation';
-import { ERROR_404 } from '../../../lib/constants/errors';
+import { ERROR_404 } from '../../../lib/constants/routes';
 
 const PostViewWrapper: React.FC = () => {
   const { postId } = useParams<{ postId: string }>();
